pyproject.toml
[project]
name = "yippie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "pandas>=2.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "yfinance>=0.2.33",
    "numpy>=2.2.6",
    "gymnasium>=0.29.1",
    "matplotlib>=3.8.0",
    "alpaca-py>=0.17.0",
    "gym-anytrading>=1.3.2",
]
